add new asset image uploading animation
